+++ CHANGELOG.md
@@ -7,6 +7,17 @@
 
 ---
 
+## v0.0.102 [unreleased]
+
+- `ryo3-jiff`
+  - fix `__radd__` impls for temporal types used to add duration types that have
+    an `__add__` method that is not aware of the ry-jiff-temporal types
+    (`datetimes.timedelta` and `ry.Duration`)
+- stubs/types
+  - fix positional dunder stubs
+
+---
+
 ## v0.0.101 [2026-09-23]
 
 - `ryo3-core`/`ryo3-serde`